Renamed UI projects into "*App"
