+++ ClientApp/ConnectWindow.da.resx
