+++ ClientApp/ConnectWindow.eo.resx
