+++ ClientApp/ConnectWindow.resx
