+++ ClientApp/Properties/AssemblyInfo.cs
