+++ ClientApp/WindowDisplay.eo.resx
