+++ ClientApp/WindowDisplay.resx
