+++ ServerApp/NativeAPI/ApplicationDropper.cs
@@ -3,7 +3,7 @@ using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Windows.Forms;
 
-namespace Server.NativeAPI;
+namespace ServerApp.NativeAPI;
 
 // https://social.msdn.microsoft.com/Forums/en-US/bfc75b57-df16-48c6-92af-ea0a34f540ae/how-to-get-the-handle-of-a-window-that-i-click?forum=csharplanguage
 internal static class ApplicationDropper
