+++ ServerApp/Properties/AssemblyInfo.cs
