+++ ServerApp/Resources/applicationDropperCursor.cur
