+++ ServerApp/Resources/applicationDropperIcon.png
