+++ ServerApp/WindowCapture.eo.resx
