+++ ServerApp/WindowCapture.resx
